requirements.txt
Babel==2.1.1
bcrypt==2.0.0
blinker==1.4
cffi==1.2.1
chardet==2.3.0
decorator==4.0.4
defusedxml==0.4.1
Flask==0.10.1
Flask-Assets==0.11
Flask-Babel==0.9
Flask-Bcrypt==0.7.1
flask-lesscss==0.9.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OAuth==0.12
Flask-OpenID==1.2.5
Flask-Scss==0.5
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
lamson==1.3.4
MarkupSafe==0.23
oauth2==1.9.0.post1
pbr==1.8.1
pycparser==2.14
pyScss==1.3.4
python3-openid==3.0.7
pytz==2015.6
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.8
sqlalchemy-migrate==0.10.0
sqlparse==0.1.16
Tempita==0.5.2
webassets==0.11.1
Werkzeug==0.10.4
wheel==0.26.0
Whoosh==2.7.0
WTForms==2.0.2
trans==2.0.2